Added merchant menu with some stat upgrades

+++ GGJ25/Assets/Project/Scripts/StatBase.cs
@@ -9,16 +9,16 @@ public class StatBase : MonoBehaviour
     [SerializeField]
     private float MovementSpeed;
     [SerializeField]
-    private int Damage;
+    private float Damage;
 
     public int GetHealth() => Health;
     public int GetMaxHealth() => MaxHealth;
     public float GetMovementSpeed() => MovementSpeed;
-    public int GetDamage() => Damage;
+    public float GetDamage() => Damage;
 
     public void SetHealth(int newHealth) => Health = newHealth;
     public void SetMaxHealth(int max) => MaxHealth = max;
-    public void SetDamage(int damage) => Damage = damage;
+    public void SetDamage(float damage) => Damage = damage;
     public void SetMovementSpeed(float movementSpeed) => MovementSpeed = movementSpeed;
     
 }
